drivers: misc: properly fix clang with the previous fix front camera crashed when deep social media option was enabled
(cherry picked from commit 2f9538aff63dff5b2d36aa150c5dcccfa27d39e4)

+++ drivers/misc/ipu/ipu-adapter-ab-mfd.c
@@ -584,8 +584,7 @@ static int ipu_adapter_ab_mfd_atomic_sync32_shared_memory(struct device *dev,
 		mutex_unlock(&dev_data->sync_lock);
 		return -ENOLINK;
 	}
-
-	if ((uintptr_t)buffer_vaddr % sizeof(uintptr_t) != 0) {
+	if ((unsigned long)buffer_vaddr % sizeof(uint32_t) != 0) {
 		dev_err(dev, "%s: error: unaligned access\n", __func__);
 		mutex_unlock(&dev_data->sync_lock);
 		return -EINVAL;
